- separate CocktailMenu enum from HomeViewModel
