+++ lib/screens/home/cocktail_menu.dart
@@ -1,0 +1,9 @@
+enum CocktailMenu {
+  categories(name: 'Category'),
+  typeOfGlass(name: 'Type of Glass'),
+  ingredient(name: 'Ingredient'),
+  alcoholic(name: 'Alcoholic/Non Alcoholic');
+
+  final String name;
+  const CocktailMenu({required this.name});
+}
